Move e2e tests to top level *.test.ts
