+++ packages/neuron-wallet/tests-e2e/application/env.ts
