--- packages/neuron-wallet/tests-e2e/tests/index.ts
@@ -1,13 +1,0 @@
-import Wallet from './wallet'
-import Network from './network'
-import SendTransaction from './send-transaction'
-import Notification from './notification'
-import AddressBook from './address-book'
-
-export default {
-  Wallet,
-  Network,
-  SendTransaction,
-  Notification,
-  AddressBook,
-}
